requirements.txt
torch
torchvision
numpy
opencv-python
pandas
pillow
dlib